test: Reenable a variety of skipped tests
